refactor(buildrequest): extract client/provider resolution into client_resolution.go
Move Concern C in full from orchestrator.go to a new
client_resolution.go (~378 LOC). Verbatim move with no behaviour change;
declarations follow the original C1-then-C2 declaration order:
ResolveProvider, ResolveEffectiveClient, ResolvePrimaryClient,
normalizeStrategyProvider, ResolveClientProvider, findRuntimeClient,
hasInvalidProviderOverride, hasInvalidStartOverride,
isInvalidOverrideReason, hasExplicitStrategyProviderWithoutStrategy,
isStrategyProvider, resolveFallbackStrategyChain.

Drops the now-unused "errors" import from orchestrator.go (only
ResolveEffectiveClient used errors.Is). Regenerates bamlutils/embed.go.

Part of #199 item 4.
